requirements.txt
flask==2.3.2 
flask-cors==3.0.10 
flask-limiter==3.4.0 
flask-wtf==1.1.1 
mysql-connector-python==8.0.33 
python-dotenv==1.0.0 
pdfplumber==0.9.0 
pymupdf==1.22.0 
pdf2image==1.16.0 
pytesseract==0.3.10 
Pillow==9.5.0 
werkzeug==2.3.6 
gunicorn==22.0.0 
python-docx==1.1.2 
python-pptx==0.6.23 
bcrypt==4.1.3 
scikit-learn==1.5.2